more updates and breaking apart resolvers
