+++ server/resolvers/auth/authMutationResolver.ts
@@ -1,65 +1,16 @@
-import { UserModel } from "../models/users";
-import { TokenModel } from "../models/token_storage";
-import { HuntModel } from "../models/hunts";
+import { UserModel } from "../../models/users";
+import { TokenModel } from "../../models/token_storage";
 import { compareSync, hashSync } from "bcryptjs";
-import { createAndSaveToken } from "../utils/jwt";
-import { returnedItems } from "../utils/transforms/returnedItems";
-import { AllUsersPayload, Resolvers } from "../generated/graphql";
-import { createBsonObjectId } from "../utils/transforms/createBsonObjectId";
+import { createAndSaveToken } from "../../utils/jwt";
+import { Resolvers } from "../../generated/graphql";
+import { createBsonObjectId } from "../../utils/transforms/createBsonObjectId";
 import {
   AuthenticationError,
   throwResolutionError,
   throwServerError,
-} from "../utils/apolloErrorHandlers";
+} from "../../utils/apolloErrorHandlers";
 
 const userResolver: Resolvers = {
-  Query: {
-    getAllUsers: async (
-      _parent: unknown,
-      _args,
-      { user },
-      { operation: { name } }
-    ) => {
-      if (!user) {
-        return AuthenticationError({
-          message: "No user stored on server.",
-          location: name?.value,
-        });
-      }
-      const users = await UserModel.find({}).select({ hash: 0 }).exec();
-      return users.map(returnedItems);
-    },
-    getUserFromToken: async (
-      _parent: unknown,
-      _args,
-      { token },
-      { operation: { name } }
-    ) => {
-      const doc = await TokenModel.findOne({ token })
-        .select({ issuedToUser: 1 })
-        .exec();
-      if (!doc) {
-        return throwResolutionError({
-          location: name?.value,
-          message: "Token does not exist.",
-        });
-      }
-
-      const user = await UserModel.findById(doc.issuedToUser).exec();
-      if (!user) {
-        return throwResolutionError({
-          location: name?.value,
-          message: "User does not exist.",
-        });
-      }
-
-      return user.transformWithoutHash();
-    },
-    userNameExists: async (_parent: unknown, { user_name }) => {
-      const matches = await UserModel.findUsername(user_name);
-      return matches.length > 0;
-    },
-  },
   Mutation: {
     registerUser: async (
       _parent: unknown,
@@ -149,14 +100,6 @@ const userResolver: Resolvers = {
           location: name?.value,
         });
       }
-    },
-  },
-  AllUsersPayload: {
-    hunts: async (parent: AllUsersPayload) => {
-      const created_by = createBsonObjectId(parent._id);
-      return await HuntModel.find({
-        created_by,
-      });
     },
   },
 };
