+++ server/resolvers/teams/teamMutationResolver.ts
@@ -1,44 +1,13 @@
-import { TeamModel } from "../models/teams";
-import { ResponseModel } from "../models/responses";
-import { Resolvers, Team } from "../generated/graphql";
-import { returnedItems } from "../utils/transforms/returnedItems";
-import { createBsonObjectId } from "../utils/transforms/createBsonObjectId";
+import { TeamModel } from "../../models/teams";
+import { Resolvers } from "../../generated/graphql";
+import { returnedItems } from "../../utils/transforms/returnedItems";
+import { createBsonObjectId } from "../../utils/transforms/createBsonObjectId";
 import {
   throwResolutionError,
   throwServerError,
-} from "../utils/apolloErrorHandlers";
+} from "../../utils/apolloErrorHandlers";
 
 const teamResolver: Resolvers = {
-  Query: {
-    getTeamsByHuntId: async (_parent: unknown, { h_id }) => {
-      const hunt_id = createBsonObjectId(h_id);
-      return await TeamModel.find({ hunt_id });
-    },
-    getTeam: async (
-      _parent: unknown,
-      { id },
-      _ctxt,
-      { operation: { name } }
-    ) => {
-      try {
-        const team = await TeamModel.findById(id);
-
-        if (!team) {
-          return throwResolutionError({
-            message: "Failed to find the specified team.",
-            location: name?.value,
-          });
-        }
-
-        return team.transformWithTypename();
-      } catch {
-        return throwServerError({
-          message: "Unable to get teams at the moment.",
-          location: name?.value,
-        });
-      }
-    },
-  },
   Mutation: {
     createSingleTeam: async (
       _parent: unknown,
@@ -148,12 +117,6 @@ const teamResolver: Resolvers = {
           location: name?.value,
         });
       }
-    },
-  },
-  Team: {
-    responses: async (parent: Team) => {
-      const team_id = createBsonObjectId(parent._id);
-      return await ResponseModel.find({ team_id });
     },
   },
 };
